move InMemoryLogBuffer into test
