+++ src/kvstore/wal/WalFileInfo.h
@@ -10,7 +10,6 @@
 #include "common/base/Base.h"
 #include "common/base/Cord.h"
 #include "kvstore/wal/Wal.h"
-#include "kvstore/wal/InMemoryLogBuffer.h"
 
 namespace nebula {
 namespace wal {
