+++ src/kvstore/wal/test/InMemoryLogBufferList.h
@@ -6,7 +6,7 @@
 
 #include "common/base/Base.h"
 #include "utils/LogIterator.h"
-#include "kvstore/wal/InMemoryLogBuffer.h"
+#include "kvstore/wal/test/InMemoryLogBuffer.h"
 
 namespace nebula {
 namespace wal {
